Rename headers and add rmm exec policy
