+++ cpp/include/internals.hpp
