+++ cpp/src/layout/exact_repulsion.hpp
