+++ cpp/src/layout/utils.hpp
